# DOT-LMS-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOT-LMS API Documentation",
        "contact": {
            "email": "ruthgetaneh5@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "DOT-LMS API Server"
        }
    ],
    "paths": {
...
        "/Grade": {
            "get": {
                "tags": [
                    "Grade"
                ],
                "summary": "Get list of Grades in the DOT-LMS",
                "description": "Returns list of Grades",
                "operationId": "575d376ca8cf55e72a8e4b394414b54d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Grade"
                ],
                "summary": "Store new Grade's data into the database",
                "operationId": "93bf7e520b7784f71d9be4a3b6be5d1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/gradebook"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/Grade/{id}": {
            "get": {
                "tags": [
                    "Grade"
                ],
                "summary": "Get a specific Grade",
                "operationId": "62ebfc4ff17b6fd6661fe2d01c53585f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Grade's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "DBUR-1255-23"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Grade"
                ],
                "summary": "Update existing Grade",
                "operationId": "71bc6199c16a6d9b3c6d1166568a68e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Grade's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "GRD-765"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/gradebook"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Grade"
                ],
                "summary": "Delete existing Grade",
                "operationId": "d175ccc1cfb0c7d9617c01b46c5ccb49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Grade's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "GRD-765"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "gradebook": {
                "title": "Grade",
                "description": "Grade data",
                "properties": {
                    "course_id": {
                        "title": "course_id",
                        "description": "The ID of the course the grade belongs to",
                        "type": "string",
                        "example": "CUS-4161"
                    },
                    "attendance": {
                        "title": "attendance",
                        "description": "The attendance value of a student",
                        "type": "string",
                        "example": "5"
                    },
                    "individual_assignment": {
                        "title": "individual_assignment",
                        "description": "individual_assignment value of the Student",
                        "type": "string",
                        "example": "15"
                    },
                    "group_assignment": {
                        "title": "group_assignment",
                        "description": "group_assignment value of the Student",
                        "type": "string",
                        "example": "15"
                    },
                    "mid_exam": {
                        "title": "mid_exam",
                        "description": " mid_exam value of the Student",
                        "type": "string",
                        "example": "15"
                    },
                    "final_exam": {
                        "title": "final_exam",
                        "description": " final_exam value of a student",
                        "type": "string",
                        "example": "50"
                    },
                    "student_id": {
                        "title": "student_id",
                        "description": " the ID of the student the grade belongs to",
                        "type": "string",
                        "example": "DBUR-4825-23"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "api_key",
                "in": "header"
            }
        }
    },
...
}